requirements.txt
flask
sqlalchemy
requests
flask_sqlalchemy
pycountry
python-dotenv
pymysql